Update algo doc and add communication model
The communication model has a problem in Spin, though I'm not sure what
it is. But the model is also flawed in other ways... I should see how
TCP handles out of order packets and such and see if it's "reasonable"
to use something like it in a stabilizing context.

A known method of communicating over a channel of arbitrary capacity $r$
which might reorder messages is to receive $r+1$ identical messages,
where $r$ is the buffer size. I would rather not do this, and this is
not what is being modeled, but it does point to why the model is flawed.