Improve protocol over UDP
The protocol is tweaked to be fast, lean (low bandwidth), and reliable.
To improve speed, we focus on streamlining actions and corruption
recovery. A quick timeout is used in such scenarios, the goal being to
force packets through any intermittent network failures.
To reduce network load, timeouts are randomized. Further, packets which
may or may not need a reply are sometimes dropped. This gives a 20-30%
decrease in traffic.
To test reliability, we simulate dropping a given percentage of packets.