Reduce Wang tileset to commutative action tileset
It's kinda like putting Conway's game of life around a circle
where cells hold finite-domain values
and the operation between them is commutative.

Step 1: Initial configuration.
Step 2: Apply operator between adjacent cells.
Step 3: The operation is disabling, so the current adjacent cells must
wait until the old ones die off. This can happen asynchronously,
provided that no cell dies before the two that spawned it.

         e                   w e v                w   v
     f       d             f       d
                          x         u          x         u
   g           c   ==>   g           c   ==>
                          y         t          y         t
     h       b             h       b
         a                   z a s                z   s

And the question equivalent to livelock detection is, for that given
operation, is there some initial configuration (which includes the
number of cells) that causes the game to run forever?
Well, seeing as we have a reduction from Wang tiles, it's undecidable!